fix: handle match fully contained in prevBuf properly
Previously, the rr.r pointer couldn't properly be reset to start reading again from the previous buffer. We now set r.rr to a negative index, if the index is still within the bounds of the prevBuf.

Also we now handle negative values of r.rr properly in the Read and ReadRune methods. If the value is negative, that indicates that we need to start reading again in the prevBuf.